make some things optional
