+++ webauthn-authenticator-rs/src/softpasskey.rs
@@ -1,7 +1,6 @@
 use crate::authenticator_hashed::AuthenticatorBackendHashedClientData;
-use crate::crypto::get_group;
+use crate::crypto::{compute_sha256, get_group};
 use crate::error::WebauthnCError;
-use crate::util::compute_sha256;
 use openssl::{bn, ec, hash, pkey, rand, sign};
 use serde_cbor::value::Value;
 use std::collections::BTreeMap;
