+++ webauthn-authenticator-rs/src/softtoken.rs
@@ -1,9 +1,8 @@
 use crate::{
     authenticator_hashed::AuthenticatorBackendHashedClientData,
-    crypto::get_group,
+    crypto::{compute_sha256, get_group},
     ctap2::commands::{value_to_vec_u8, GetInfoResponse},
     error::WebauthnCError,
-    util::compute_sha256,
 };
 use openssl::x509::{
     extension::{AuthorityKeyIdentifier, BasicConstraints, KeyUsage, SubjectKeyIdentifier},
@@ -830,15 +829,32 @@ impl AuthenticatorBackendHashedClientData for SoftTokenFile {
 }
 
 #[cfg(test)]
+#[allow(clippy::panic)]
 mod tests {
     use super::*;
-    use crate::prelude::{Url, WebauthnAuthenticator};
+    use openssl::{hash::MessageDigest, rand::rand_bytes, sign::Verifier, x509::X509};
     use std::collections::BTreeSet;
     use tempfile::tempfile;
-    use webauthn_rs_core::proto::{AttestationCa, AttestationCaList};
-    use webauthn_rs_core::WebauthnCore as Webauthn;
+    use webauthn_rs_core::{
+        proto::{AttestationCa, AttestationCaList, COSEKey},
+        WebauthnCore as Webauthn,
+    };
     use webauthn_rs_proto::{
-        AttestationConveyancePreference, COSEAlgorithm, UserVerificationPolicy,
+        AllowCredentials, AttestationConveyancePreference, COSEAlgorithm, PubKeyCredParams,
+        RelyingParty, User, UserVerificationPolicy,
+    };
+
+    use crate::{
+        ctap2::{
+            commands::{
+                value_to_vec_u8, GetAssertionRequest, GetAssertionResponse, MakeCredentialRequest,
+                MakeCredentialResponse,
+            },
+            CBORResponse,
+        },
+        perform_auth_with_request, perform_register_with_request,
+        prelude::{Url, WebauthnAuthenticator},
+        softtoken::SoftToken,
     };
 
     #[test]
@@ -1014,5 +1030,147 @@ mod tests {
             .authenticate_credential(&r, &auth_state)
             .expect("webauth authentication denied");
         info!("auth_res -> {:x?}", auth_res);
+    }
+
+    #[test]
+    fn perform_register_auth_with_command() {
+        let _ = tracing_subscriber::fmt::try_init();
+        let (mut soft_token, _) = SoftToken::new().unwrap();
+        let mut client_data_hash = vec![0; 32];
+        let mut user_id = vec![0; 16];
+        rand_bytes(&mut client_data_hash).unwrap();
+        rand_bytes(&mut user_id).unwrap();
+
+        let request = MakeCredentialRequest {
+            client_data_hash: client_data_hash.clone(),
+            rp: RelyingParty {
+                name: "example.com".to_string(),
+                id: "example.com".to_string(),
+            },
+            user: User {
+                id: Base64UrlSafeData(user_id),
+                name: "sampleuser".to_string(),
+                display_name: "Sample User".to_string(),
+            },
+            pub_key_cred_params: vec![
+                PubKeyCredParams {
+                    type_: "public-key".to_string(),
+                    alg: -7,
+                },
+                PubKeyCredParams {
+                    type_: "public-key".to_string(),
+                    alg: -257,
+                },
+            ],
+            exclude_list: vec![],
+            options: None,
+            pin_uv_auth_param: None,
+            pin_uv_auth_proto: None,
+            enterprise_attest: None,
+        };
+
+        let response = perform_register_with_request(&mut soft_token, request, 10000).unwrap();
+
+        // All keys should be ints
+        let m: Value = serde_cbor::from_slice(response.as_slice()).unwrap();
+        let m = if let Value::Map(m) = m {
+            m
+        } else {
+            panic!("unexpected type")
+        };
+        assert!(m.keys().all(|k| matches!(k, Value::Integer(_))));
+
+        // Try to deserialise the MakeCredentialResponse again
+        let response =
+            <MakeCredentialResponse as CBORResponse>::try_from(response.as_slice()).unwrap();
+        trace!(?response);
+
+        // Run packed attestation verification
+        // https://www.w3.org/TR/webauthn-2/#sctn-packed-attestation
+        let mut att_stmt = if let Value::Map(m) = response.att_stmt.unwrap() {
+            m
+        } else {
+            panic!("unexpected type");
+        };
+        trace!(?att_stmt);
+        let signature = value_to_vec_u8(
+            att_stmt.remove(&Value::Text("sig".to_string())).unwrap(),
+            "att_stmt.sig",
+        )
+        .unwrap();
+
+        // Extract attestation certificate
+        let x5c = if let Value::Array(v) = att_stmt.remove(&Value::Text("x5c".to_string())).unwrap()
+        {
+            v
+        } else {
+            panic!("Unexpected type");
+        };
+        let x5c = value_to_vec_u8(x5c[0].to_owned(), "x5c[0]").unwrap();
+        let verification_cert = X509::from_der(&x5c).unwrap();
+        let pubkey = verification_cert.public_key().unwrap();
+
+        // Reconstruct verification data (auth_data + client_data_hash)
+        let mut verification_data =
+            value_to_vec_u8(response.auth_data.unwrap(), "verification_data").unwrap();
+        let auth_data_len = verification_data.len();
+        verification_data.reserve(client_data_hash.len());
+        verification_data.extend_from_slice(&client_data_hash);
+
+        let mut verifier = Verifier::new(MessageDigest::sha256(), &pubkey).unwrap();
+        assert!(verifier
+            .verify_oneshot(&signature, &verification_data)
+            .unwrap());
+
+        // https://www.w3.org/TR/webauthn-2/#attestation-object
+        let cred_id_off = /* rp_id_hash */ 32 + /* flags */ 1 + /* counter */ 4 + /* aaguid */ 16;
+        let cred_id_len = u16::from_be_bytes(
+            (&verification_data[cred_id_off..cred_id_off + 2])
+                .try_into()
+                .unwrap(),
+        ) as usize;
+        let cred_id = Base64UrlSafeData(
+            (verification_data[cred_id_off + 2..cred_id_off + 2 + cred_id_len]).to_vec(),
+        );
+
+        // Future assertions are signed with this COSEKey
+        let cose_key: Value = serde_cbor::from_slice(
+            &verification_data[cred_id_off + 2 + cred_id_len..auth_data_len],
+        )
+        .unwrap();
+        let cose_key = COSEKey::try_from(&cose_key).unwrap();
+
+        rand_bytes(&mut client_data_hash).unwrap();
+        let request = GetAssertionRequest {
+            client_data_hash: client_data_hash.clone(),
+            rp_id: "example.com".to_string(),
+            allow_list: vec![AllowCredentials {
+                type_: "public-key".to_string(),
+                id: cred_id.to_owned(),
+                transports: None,
+            }],
+            options: None,
+            pin_uv_auth_param: None,
+            pin_uv_auth_proto: None,
+        };
+        trace!(?request);
+
+        let response = perform_auth_with_request(&mut soft_token, request, 10000).unwrap();
+        let response =
+            <GetAssertionResponse as CBORResponse>::try_from(response.as_slice()).unwrap();
+        trace!(?response);
+
+        // Check correct matching credential
+        assert_eq!(response.credential.unwrap().id, cred_id);
+
+        // Check the signature
+        let signature = response.signature.unwrap();
+        let mut verification_data = response.auth_data.unwrap();
+        verification_data.reserve(client_data_hash.len());
+        verification_data.extend_from_slice(&client_data_hash);
+
+        assert!(cose_key
+            .verify_signature(&signature, &verification_data)
+            .unwrap());
     }
 }
